Refactor EM sampling: move _em_sample_filter_sql to em_sampling module and update imports in related files
